debt: server/auth: extract user-handlers.js from webauthn.js
Move user + passkey CRUD endpoints (list, create, update, delete
+ passkey update/delete) out of webauthn.js. Inlines small
requireUser/requireAdmin/jsonResponse/serializeCredential helpers
rather than importing back from webauthn.js to avoid circular
require.

webauthn.js: 719 → 525 lines. Stays as the WebAuthn ceremony
module + request router + invitation endpoints.

knip.json: add user-handlers.js as an entry — its exports are
consumed via property access on the required module (knip's CJS
analysis can't trace that).

npm test:unit 782/782 pass. knip clean.
